# Conflicts:
#	analysis/analysis.iml
#	chat/chat.iml
#	chat/src/main/resources/application.yaml
#	frontend/frontend.iml
#	indent/indent.iml
#	indent/pom.xml
#	indent/src/main/resources/application.yaml
#	indent/src/test/java/com/cross2u/indent/IndentApplicationTests.java
#	manage/manage.iml
#	manage/src/main/resources/application.yaml
#	pay/pay.iml
#	search/search.iml
#	server/server.iml
#	store/src/main/java/com/cross2u/store/StoreApplication.java
#	store/src/main/java/com/cross2u/store/model/base/BaseManufacturer.java
#	store/store.iml
#	user/pom.xml
#	user/src/main/java/com/cross2u/user/UserApplication.java
#	user/src/main/java/com/cross2u/user/controller/AdminControllerL.java
#	user/src/main/java/com/cross2u/user/model/base/BaseAbnormalminfo.java
#	user/src/test/java/com/cross2u/user/UserApplicationTests.java
#	ware/src/main/java/com/cross2u/ware/WareApplication.java
#	ware/src/main/java/com/cross2u/ware/service/WareServiceL.java
#	ware/src/test/java/com/cross2u/ware/WareApplicationTests.java

+++ ware/src/main/java/com/cross2u/ware/model/base/BaseWare.java
@@ -177,7 +177,7 @@ public abstract class BaseWare<M extends BaseWare<M>> extends Model<M> implement
 		return get("wDeliverHour");
 	}
 
-	public void setWStartNum(Long wStartNum) {
+	public void setWStartNum(Integer wStartNum) {
 		set("wStartNum", wStartNum);
 	}
 
@@ -185,7 +185,7 @@ public abstract class BaseWare<M extends BaseWare<M>> extends Model<M> implement
 		return get("wStartNum");
 	}
 
-	public void setWHighNum(Long wHighNum) {
+	public void setWHighNum(Integer wHighNum) {
 		set("wHighNum", wHighNum);
 	}
